requirements.txt
#devops
black==25.1.0
click==8.1.3 
pytest==8.3.4
pytest-cov==6.0.0
pylint==3.3.4
sentence_transformers==3.4.1
ollama
ipykernel
openai==1.62.0
python-dotenv
pandas
numpy
langchain_neo4j 
transformers
langchain
langchain-community 
langchain-ollama 
langchain-experimental 
neo4j 
tiktoken 
yfiles_jupyter_graphs 
json-repair 
langchain-openai 
langchain_core
gspread
oauth2client
nltk
rouge_score
logging
matplotlib
seaborn